+++ sdk/src/com/appnexus/opensdk/MediatedNativeAdController.java
@@ -224,7 +224,7 @@ public class MediatedNativeAdController {
 
     }
 
-    private class ResultCBRequest extends HTTPGet<Void, Void, HTTPResponse> {
+    private class ResultCBRequest extends HTTPGet {
         WeakReference<AdRequester> requester;
         private final String resultCB;
         final ResultCode result;